spider/requirements.txt
requests==2.32.3
mysql-connector-python==9.1.0
kafka-python==2.0.2
